refactor: integrate adminjs and relocate AR app
