+++ AGENTS.md
@@ -54,9 +54,8 @@ project-root/
 │   ├── marker.png         # PNG изображение маркера (можно несколько)
 │   ├── target.mind        # Сгенерированный файл таргета (для каждого маркера свой .mind)
 │   └── assets/            # Прочие ассеты (3D-модели, изображения)
-├── src/
-│   ├── ar-scene.js        # Инициализация и логика AR
-│   └── utils/             # Вспомогательные функции (по необходимости)
+├── ar-app/
+│   └── src/               # код AR и утилиты
 ├── server.js             # API сервер на Express
 ├── tests/                # тесты Vitest
 ├── index.html
@@ -137,7 +136,7 @@ pnpm preview  # проверка dist/
    - Загрузите свой маркер (PNG/JPG)
    - Скачай target.mind и положи в `public/` (рядом с marker.png)
 3. Добавь новые ассеты (3D-модели, изображения) в `public/assets/`.
-4. Пиши основную логику в `src/ar-scene.js`; все ассеты загружаются из `public/`.
+4. Пиши основную логику в `ar-app/src/ar-scene.js`; все ассеты загружаются из `public/`.
 5. Настрой кодстайл и автоформатирование (см. [Настройка VS Code](#vs-code-setup)).
 6. Для переменных окружения используй `.env` (пример — `.env.example`).
    - Для отправки аналитики укажи `VITE_ANALYTICS_ENDPOINT=<url>`
