+++ ar-app/src/register.js
