+++ ar-app/src/utils/auth.js
