+++ cms/app/catalog.html
@@ -51,6 +51,6 @@
   <script src="../majestic/vendor.bundle.base.js"></script>
   <script src="../js/jquery.cookie.js"></script>
   <script src="../majestic/template.js"></script>
-  <script type="module" src="../../src/catalog.js"></script>
+  <script type="module" src="../../ar-app/src/catalog.js"></script>
 </body>
 </html>
