+++ cms/app/login.html
@@ -45,6 +45,6 @@
   <script src="../majestic/vendor.bundle.base.js"></script>
   <script src="../js/jquery.cookie.js"></script>
   <script src="../majestic/template.js"></script>
-  <script type="module" src="../../src/login.js"></script>
+  <script type="module" src="../../ar-app/src/login.js"></script>
 </body>
 </html>
