+++ cms/app/profile.html
@@ -67,6 +67,6 @@
   <script src="../majestic/vendor.bundle.base.js"></script>
   <script src="../js/jquery.cookie.js"></script>
   <script src="../majestic/template.js"></script>
-  <script type="module" src="../../src/profile.js"></script>
+  <script type="module" src="../../ar-app/src/profile.js"></script>
 </body>
 </html>
