+++ cms/app/register.html
@@ -48,6 +48,6 @@
   <script src="../majestic/vendor.bundle.base.js"></script>
   <script src="../js/jquery.cookie.js"></script>
   <script src="../majestic/template.js"></script>
-  <script type="module" src="../../src/register.js"></script>
+  <script type="module" src="../../ar-app/src/register.js"></script>
 </body>
 </html>
