+++ cms/index-old.html
@@ -145,6 +145,6 @@
     </div>
     <script src="./majestic/vendor.bundle.base.js"></script>
     <script src="./majestic/template.js"></script>
-    <script type="module" src="../src/cp.js"></script>
+    <script type="module" src="../ar-app/src/cp.js"></script>
   </body>
 </html>
