+++ cms/index.html
@@ -643,7 +643,7 @@
   <!-- End custom js for this page-->
 
   <script src="js/jquery.cookie.js" type="text/javascript"></script>
-  <script type="module" src="../src/cp.js"></script>
+  <script type="module" src="../ar-app/src/cp.js"></script>
   <script type="module">
     const base = import.meta.env.BASE_URL;
     document
