+++ package.json
@@ -21,8 +21,11 @@
     "check-api": "node scripts/check-api.js"
   },
   "dependencies": {
+    "@adminjs/express": "^6.1.1",
+    "@adminjs/mongoose": "^4.1.0",
     "@aws-sdk/client-s3": "^3.548.0",
     "@aws-sdk/s3-request-presigner": "^3.548.0",
+    "adminjs": "^7.8.16",
     "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
     "dotenv": "^16.5.0",
