+++ server.js
@@ -20,10 +20,22 @@ import path from 'node:path';
 import fs from 'node:fs';
 import jwt from 'jsonwebtoken';
 import rateLimit from 'express-rate-limit';
+import AdminJS from 'adminjs';
+import AdminJSExpress from '@adminjs/express';
+import * as AdminJSMongoose from '@adminjs/mongoose';
 
 export const app = express();
 app.use(express.json());
 app.use(helmet());
+
+AdminJS.registerAdapter(AdminJSMongoose);
+
+const admin = new AdminJS({
+  databases: [mongoose],
+  rootPath: '/admin',
+});
+const adminRouter = AdminJSExpress.buildRouter(admin);
+app.use(admin.options.rootPath, adminRouter);
 const allowedOrigins = process.env.FRONTEND_ORIGINS
   ? process.env.FRONTEND_ORIGINS.split(',')
       .map((o) => o.trim())
