+++ tailwind.config.js
@@ -1,6 +1,6 @@
 export default {
   darkMode: 'class',
-  content: ['./index.html', './cms/index.html', './src/**/*.{js,html}'],
+  content: ['./index.html', './cms/index.html', './ar-app/src/**/*.{js,html}'],
   theme: {
     extend: {},
   },
