+++ vite.config.js
@@ -1,5 +1,5 @@
 import { defineConfig, loadEnv } from 'vite';
-import postcss from './src/postcss.config.js';
+import postcss from './ar-app/src/postcss.config.js';
 import { viteStaticCopy } from 'vite-plugin-static-copy';
 
 export default defineConfig(({ mode }) => {
